requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
Automat==24.8.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
decorator==5.1.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozenlist==1.4.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
multidict==6.0.5
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.4
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
tornado==6.4.1
trio==0.26.2
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.9.11
zope.interface==7.0.3
